# 8270947: AArch64: C1: use zero_words to initialize all objects