--- a/src/hotspot/cpu/aarch64/macroAssembler_aarch64.cpp
+++ b/src/hotspot/cpu/aarch64/macroAssembler_aarch64.cpp
@@ -4680,10 +4680,13 @@ const int MacroAssembler::zero_words_block_size = 8;
 address MacroAssembler::zero_words(Register ptr, Register cnt)
 {
   assert(is_power_of_2(zero_words_block_size), "adjust this");
-  assert(ptr == r10 && cnt == r11, "mismatch in register usage");
 
   BLOCK_COMMENT("zero_words {");
-  cmp(cnt, (u1)zero_words_block_size);
+  assert(ptr == r10 && cnt == r11, "mismatch in register usage");
+  RuntimeAddress zero_blocks = RuntimeAddress(StubRoutines::aarch64::zero_blocks());
+  assert(zero_blocks.target() != NULL, "zero_blocks stub has not been generated");
+
+  subs(rscratch1, cnt, zero_words_block_size);
   Label around;
   br(LO, around);
   {
@@ -4701,6 +4704,9 @@ address MacroAssembler::zero_words(Register ptr, Register cnt)
     }
   }
   bind(around);
+
+  // We have a few words left to do. zero_blocks has adjusted r10 and r11
+  // for us.
   for (int i = zero_words_block_size >> 1; i > 1; i >>= 1) {
     Label l;
     tbz(cnt, exact_log2(i), l);
@@ -4715,55 +4721,29 @@ address MacroAssembler::zero_words(Register ptr, Register cnt)
     str(zr, Address(ptr));
     bind(l);
   }
+
   BLOCK_COMMENT("} zero_words");
-  postcond(pc() != badAddress);
   return pc();
 }
 
 // base:         Address of a buffer to be zeroed, 8 bytes aligned.
 // cnt:          Immediate count in HeapWords.
+//
+// r10, r11, rscratch1, and rscratch2 are clobbered.
 #define SmallArraySize (18 * BytesPerLong)
 void MacroAssembler::zero_words(Register base, uint64_t cnt)
 {
   STATIC_ASSERT(zero_words_block_size < SmallArraySize);
-  BLOCK_COMMENT("zero_words {");
-
   if (cnt <= SmallArraySize / BytesPerLong) {
     int i = cnt & 1;  // store any odd word to start
     if (i) str(zr, Address(base));
-
     for (; i < (int)cnt; i += 2) {
       stp(zr, zr, Address(base, i * wordSize));
     }
-
   } else {
-    push(RegSet::of(r10, r11), sp);
-
     mov(r10, base); mov(r11, cnt);
-    RuntimeAddress zero_blocks = RuntimeAddress(StubRoutines::aarch64::zero_blocks());
-    assert(zero_blocks.target() != NULL, "zero_blocks stub has not been generated");
-
-    address tpc = trampoline_call(zero_blocks);
-    if (tpc == NULL) {
-      postcond(pc() == badAddress);
-      return;
-    }
-
-    // We have a few words left to do. zero_blocks has adjusted r10
-    // for us.
-    cnt %= zero_words_block_size;
-    int i = cnt & 1;  // store any odd word to start
-    if (i) str(zr, Address(r10));
-
-    for (; i < (int)cnt; i += 2) {
-      stp(zr, zr, Address(r10, i * wordSize));
-    }
-
-    pop(RegSet::of(r10, r11), sp);
+    zero_words(r10, r11);
   }
-
-  BLOCK_COMMENT("} zero_words");
-  postcond(pc() != badAddress);
 }
 
 // Zero blocks of memory by using DC ZVA.