@nick-arm 2021-07-29
I hit the following assertion failure when I ran tier1 with `-XX:TieredStopAtLevel=1`. This was from `compiler/loopopts/TestDeepGraphVerifyIterativeGVN.java`.

```
#  Internal Error (/mnt/nicgas01-pc/jdk/src/hotspot/cpu/aarch64/macroAssembler_aarch64.cpp:566), pid=124784, tid=124797
#  assert(pc() == ((address)::badAddressVal)) failed: postcond
```

```
V  [libjvm.so+0x12c1bbc]  MacroAssembler::trampoline_call(Address, CodeBuffer*)+0x2fc
V  [libjvm.so+0x12daf10]  MacroAssembler::zero_words(RegisterImpl*, RegisterImpl*)+0x130
V  [libjvm.so+0x79c950]  C1_MacroAssembler::initialize_body(RegisterImpl*, RegisterImpl*, int, RegisterImpl*, RegisterImpl*)+0x1b0
V  [libjvm.so+0x79daec]  C1_MacroAssembler::allocate_array(RegisterImpl*, RegisterImpl*, RegisterImpl*, RegisterImpl*, int, int, RegisterImpl*, Label&)+0x41c
V  [libjvm.so+0x73bad0]  LIR_Assembler::emit_alloc_array(LIR_OpAllocArray*)+0x160
V  [libjvm.so+0x71be40]  LIR_OpAllocArray::emit_code(LIR_Assembler*)+0x20
V  [libjvm.so+0x72c6e8]  LIR_Assembler::emit_lir_list(LIR_List*)+0xe8
V  [libjvm.so+0x72ce48]  LIR_Assembler::emit_code(BlockList*)+0x78
V  [libjvm.so+0x6d394c]  Compilation::emit_code_body()+0x14c
V  [libjvm.so+0x6d40a0]  Compilation::compile_java_method()+0x550
```