@nick-arm 2021-07-29
> I hit the following assertion failure when I ran tier1 with `-XX:TieredStopAtLevel=1`. This was from `compiler/loopopts/TestDeepGraphVerifyIterativeGVN.java`.
> 
> ```
> #  Internal Error (/mnt/nicgas01-pc/jdk/src/hotspot/cpu/aarch64/macroAssembler_aarch64.cpp:566), pid=124784, tid=124797
> #  assert(pc() == ((address)::badAddressVal)) failed: postcond
> ```

Maybe we need to reserve some extra space for the trampoline stubs in C1's `Compilation::setup_code_buffer()`? That assert seems dodgy though, `pc()` won't be `badAddressVal` if `code()->blob()` was initially NULL (the code buffer wasn't resizeable).