@adinn 2021-07-30
My initial reaction to that extra last commit was that anything other than a compiler thread calling MacroAssembler::zero_words implies "Something is rotten in the state of Danmark" and the only end game is a stage strewn with corpses. Of course, zero_words gets called from the stub generator. However, I guess the stub complete check avoids any problems there.

Were you thinking of some other potential use for it? e.g. lazy intrinsics?

Anyway, the check certainly does no harm so it is all still good.